+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - detailed logger
+### Changed
+- update gems
+- disable history tests
+- enable announcements tests
 
 ## [0.1.3] - 2019-07-02
 ### Changed